+++ bundle/src/main/java/com/adobe/acs/commons/mcp/impl/processes/asset/AssetIngestor.java
@@ -135,7 +135,8 @@ public abstract class AssetIngestor extends ProcessDefinition {
     String jcrBasePath = "/content/dam";
     @FormField(
             name = "Ignore folders",
-            description = "List of folder names to be ignored",
+            description = "List of folder names to be ignored. If you want to exclude folder name add '-' sign before name."
+                          + "If you want to include name, just write folder name or add '+' sign before name",
             hint = "tmp,.DS_STORE",
             options = {"default=-tmp,-ds_store,-.ds_store,-.thumbs,-.appledouble"}
     )
@@ -144,7 +145,8 @@ public abstract class AssetIngestor extends ProcessDefinition {
 
     @FormField(
             name = "Ignore files",
-            description = "List of file names to ignore",
+            description = "List of file names to ignore If you want to exclude file name add '-' sign before name."
+                          + "If you want to include name, just write file name or add '+' sign before name",
             hint = "full file names, comma separated",
             options = {"default=-ds_store,-.ds_store"}
     )
@@ -153,7 +155,8 @@ public abstract class AssetIngestor extends ProcessDefinition {
 
     @FormField(
             name = "Ignore extensions",
-            description = "List of file extensions to ignore",
+            description = "List of file extensions to ignore. If you want to exclude extension add '-' sign before name."
+                          + "If you want to include extension, just write extension or add '+' sign before name",
             hint = "mp4,txt, etc.",
             options = {"default=-txt,-html,-css,-js,-thm,-exe,-db"}
     )