+++ bundle/src/main/java/com/adobe/acs/commons/mcp/impl/processes/asset/NamesFilter.java
@@ -1,10 +1,13 @@
 package com.adobe.acs.commons.mcp.impl.processes.asset;
 
 import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 class NamesFilter {
 
@@ -15,7 +18,11 @@ class NamesFilter {
     }
 
     NamesFilter(String filterString) {
-        List<String> filterExpressions = Arrays.asList(filterString.trim().toLowerCase().split(","));
+        List<String> filterExpressions = Optional.ofNullable(filterString)
+                                                 .filter(StringUtils::isNotEmpty)
+                                                 .map(str -> str.trim().toLowerCase().split(","))
+                                                 .map(Arrays::asList)
+                                                 .orElse(Collections.emptyList());
 
         for (String filterExpression : filterExpressions) {
             if (filterExpression.startsWith("+")) {